this is a much better format actually
